Rework data importing from other plugin.
