+++ src/main/java/com/onarandombox/multiverseinventories/dataimport/multiinv/MIInventoryOldWrapper.java
@@ -1,14 +1,14 @@
-package com.onarandombox.multiverseinventories.migration.multiinv;
+package com.onarandombox.multiverseinventories.dataimport.multiinv;
 
 import org.bukkit.inventory.ItemStack;
-import uk.co.tggl.pluckerpluck.multiinv.inventory.MIInventory;
+import uk.co.tggl.pluckerpluck.multiinv.inventory.MIInventoryOld;
 
 /**
- * Wraps MIInventory to provide a way of accessing the inventory/armor contents.
+ * Wraps MIInventoryOld to provide a way of accessing the inventory/armor contents.
  */
-public class MIInventoryWrapper extends MIInventory implements MIInventoryInterface {
+public class MIInventoryOldWrapper extends MIInventoryOld implements MIInventoryInterface {
 
-    public MIInventoryWrapper(String inventoryString) {
+    public MIInventoryOldWrapper(String inventoryString) {
         super(inventoryString);
     }
