+++ src/main/java/com/onarandombox/multiverseinventories/dataimport/multiinv/package-info.java
@@ -1,5 +1,4 @@
 /**
  * This package contains MultiInv classes to handle importing their data.
  */
-package com.onarandombox.multiverseinventories.migration.multiinv;
-
+package com.onarandombox.multiverseinventories.dataimport.multiinv;
