+++ src/main/java/com/onarandombox/multiverseinventories/dataimport/package-info.java
@@ -1,0 +1,5 @@
+/**
+ * This package contains importer classes to help with importing player stats/inventory data from
+ * other similar plugins into Multiverse Inventories.
+ */
+package com.onarandombox.multiverseinventories.dataimport;
