+++ src/main/java/com/onarandombox/multiverseinventories/dataimport/worldinventories/package-info.java
@@ -1,5 +1,4 @@
 /**
  * This package contains WorldInventories classes to handle importing their data.
  */
-package com.onarandombox.multiverseinventories.migration.worldinventories;
-
+package com.onarandombox.multiverseinventories.dataimport.worldinventories;
