--- src/main/java/com/onarandombox/multiverseinventories/migration/DataImporter.java
@@ -1,22 +1,0 @@
-package com.onarandombox.multiverseinventories.migration;
-
-import org.bukkit.plugin.Plugin;
-
-/**
- * Interface for data migration importers.
- */
-public interface DataImporter {
-
-    /**
-     * Imports the data from another plugin.
-     *
-     * @throws MigrationException If there was any MAJOR issue loading the data.
-     */
-    void importData() throws MigrationException;
-
-    /**
-     * @return The plugin associated with this Importer.
-     */
-    Plugin getPlugin();
-}
-
