--- src/main/java/com/onarandombox/multiverseinventories/migration/package-info.java
@@ -1,6 +1,0 @@
-/**
- * This package contains thigns to help with importing player stats/inventory data from
- * other similar plugins.
- */
-package com.onarandombox.multiverseinventories.migration;
-
