+++ src/main/resources/en.yml
@@ -21,7 +21,7 @@ ERROR_DATA_LOAD: 'Encountered an error while loading the data file.  Disabling..
 ERROR_NO_GROUP: '&6There is no group with the name: &f%1'
 ERROR_NO_WORLD: '&6There is no world with the name: &f%1'
 ERROR_NO_WORLD_PROFILE: '&6There is no world profile for the world: &f%1'
-ERROR_PLUGIN_NOT_ENABLED: '&f%1 &6is not enabled so you may not import data from it!'
+ERROR_PLUGIN_NOT_ENABLED: '&f%1 &6is not enabled on your server, so you may not import data from it!'
 ERROR_UNSUPPORTED_IMPORT: '&6Sorry, ''&f%1&6'' is not supported for importing.'
 ERROR_NO_SHARES_SPECIFIED: '&cYou did not specify any valid shares!'
 
@@ -87,3 +87,8 @@ DEBUG_SET:
 NOW_USING_OPTIONAL: '&f%1 &6will now be considered when player''s change world.'
 NOW_NOT_USING_OPTIONAL: '&f%1 &6will no longer be considered when player''s change world.'
 NO_OPTIONAL_SHARES: '&f%1 &6is not an optional share!'
+
+# Import command
+IMPORT_ATTEMPT: '&6Attempting to import data from %1...'
+IMPORT_FAILED: '&cAn errored occurred while attempting to import from %1.'
+IMPORT_SUCCESSFUL: '&aImported data from %1! You can now remove &6%1 &aand restart your server. Enjoy!'
